requirements.txt
openai-whisper==20.5.0
tqdm==4.66.1
ffmpeg-python==0.2.0